Back-End/prog2/pyproject.toml
[tool.poetry]
name = "prog2"
version = "0.1.0"
description = ""
authors = ["forthecat <kitekite0928@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "~4.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"
python3-nmap = "^1.6.0"
requests = "^2.31.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"
sentry-sdk = "^1.35.0"
mysqlclient = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
